Array trait takes a numpy dtype object, not a trait, as argument

+++ simphony_ui/openfoam_boundary_conditions.py
@@ -1,3 +1,4 @@
+import numpy as np
 from traits.api import HasStrictTraits, Float, Enum, Instance, Array
 from traitsui.api import View, Item, Tabbed, UItem, VGroup
 
@@ -8,7 +9,7 @@ class BoundaryConditionModel(HasStrictTraits):
     type = Enum('none', 'empty', 'zeroGradient', 'fixedGradient', 'fixedValue')
 
     #: A fixed value of velocity gradient for the boundary
-    fixed_gradient = Array(Float, (1, 3))
+    fixed_gradient = Array(np.float, (1, 3))
 
     traits_view = View(
         VGroup(
@@ -29,7 +30,7 @@ class BoundaryConditionModel(HasStrictTraits):
 class VelocityBoundaryConditionModel(BoundaryConditionModel):
 
     #: A fixed value of velocity for the boundary
-    fixed_value = Array(Float, (1, 3))
+    fixed_value = Array(np.float, (1, 3))
 
 
 class PressureBoundaryConditionModel(BoundaryConditionModel):
